### 2017-05-18 #263 @whedon
[PRE REVIEW]: Windrose
**Submitting author:** @scls19fr (<a href="http://orcid.org/0000-0001-9987-4338">Sébastien Celles</a>)
**Repository:** <a href="https://github.com/python-windrose/windrose/" target ="_blank">https://github.com/python-windrose/windrose/</a>
**Version:** 1.6
