### 2017-05-23 #263 @arfon
@tacaswell - thanks for volunteering to review this and #267. Please don't feel like you need to do these in parallel!!
