### 2017-05-24 b45068c
Add calm limit parameter and representative circle
