### 2017-07-26 #268 @tacaswell
- I am the lead developer of a dependency (Matplotlib), but I do not think that is a conflict of interest.
 - This is licensed under "CeCILL-B FREE SOFTWARE LICENSE AGREEMENT" which is labeled at the bottom as v1.0.  There is a CeCILL license on the OSI list is "Cea Cnrs Inria Logiciel Libre License, version 2.1 (CECILL-2.1)"
 - there is not a 'github release' but the version on pypi matches 1.6
 - this is a fork / continuation of an existing project.  The original author is clearly cited and included in the paper, have they been consulted?
 - are there docs other than the README?  If so I could not easily find them.
 - paper has no references, are any required?
