### 2017-07-26 #268 @scls19fr
> there is not a 'github release' but the version on pypi matches 1.6

Should we add a tag on GitHub? If that's necessary we can do it.

> - this is a fork / continuation of an existing project. The original author is clearly cited and included in the paper, have they been consulted?

@LionelR have been consulted see https://github.com/python-windrose/windrose/issues/54

> are there docs other than the README? If so I could not easily find them.

No other docs than the README

> paper has no references, are any required?

I don't really know what should be part of this field (I've never published such paper).
