### 2017-07-28 #268 @tacaswell
Yes, pushing tags that match at least the most recent version is (I think) required.  The tags are what are archived, give you a nice single-file source download, and you can use zenodo.org to get a DOI for the tag / release.
