### 2017-07-28 b287da9
Fix CI for Python 2.7 3.4 3.5 3.6 PEP8 Remove Python 3.3 support
