+++ samples/example_subplots.py
@@ -74,5 +74,6 @@ def main(filename, year):
 
     plt.show()
 
+
 if __name__ == "__main__":
     main()
