### 2017-07-30 1337cd4
Update README.md
DOI (from Zenodo)
