### 2017-07-30 #268 @scls19fr
I've just release version 1.6.1 (Zenodo DOI https://zenodo.org/record/836356#.WX2aPdOLRZ0 , tag on GitHub, upload on PyPi https://pypi.python.org/pypi?%3Aaction=pkg_edit&name=windrose )

My release procedure is defined at https://github.com/python-windrose/windrose/issues/59
