### 2017-07-30 #268 @tacaswell
@arfon Who should be consulted about the license issue?
