### 2017-07-30 python-windrose/windrose#65 @scls19fr
Inconsistent licence files
Following https://github.com/openjournals/joss-reviews/issues/268#issuecomment-318916583

In
https://github.com/python-windrose/windrose/blob/master/setup.py
it's `License :: OSI Approved :: CEA CNRS Inria Logiciel Libre License, version 2.1 (CeCILL-2.1)'`

but in https://github.com/python-windrose/windrose/blob/master/LICENCE
and in https://github.com/python-windrose/windrose/blob/master/windrose/version.py#L5
it's CeCILL-B

In https://pypi.python.org/pypi?%3Aaction=list_classifiers
CeCILL-2.1 is the only CeCILL license (moreover it's OSI approved)
so if we change to use CeCILL-B everywhere in this project we need to know what should be written in `setup.py` (maybe only its [SPDX](https://en.wikipedia.org/wiki/SPDX): `CeCILL-B`)

@LionelR Should we update update `setup.py` or `LICENSE` (and `version.py`)?

I noticed that some other projects have same "problem".
https://www.google.fr/search?q=Cecill-b+python+setup.py
https://github.com/d9pouces/django-floor
https://github.com/akaszynski/PyACVD
...

CeCILL-2.1 is a GPL compatible french license
CeCILL-B License is a BSD compatible french license (GPL incompatible)

https://fr.wikipedia.org/wiki/Licence_CeCILL

CeCILL-B imposes author(s) citation not CeCILL-2.1
