### 2017-07-30 python-windrose/windrose#65 @scls19fr
Original version from http://youarealegend.blogspot.com/search/label/windrose was CeCILL-B.
I did a mistake when I created `setup.py` and this repository. I put the only CeCILL license available
in https://pypi.python.org/pypi?%3Aaction=list_classifiers which was CeCILL-2.1
