### 2017-08-02 #268 @arfon
@tacaswell - it seems like the authors are discussing how to resolve this in https://github.com/python-windrose/windrose/issues/65 ?
