### 2017-08-02 python-windrose/windrose#65 @LionelR
@scls19fr please keep the CeCILL-B license wherever it's necessary
