+++ windrose/windrose.py
@@ -39,7 +39,7 @@ class WindAxesFactory(object):
                 ax = cls.from_ax(ax, *args, **kwargs)
                 return ax
         else:
-            raise(NotImplementedError("typ=%r but it might be in %s" % (typ, d.keys())))
+            raise (NotImplementedError("typ=%r but it might be in %s" % (typ, d.keys())))
 
 
 class WindroseAxes(PolarAxes):
@@ -129,7 +129,7 @@ class WindroseAxes(PolarAxes):
         if not self.rmax:
             self.rmax = np.max(np.sum(self._info['table'], axis=0))
         calm_count = self.calm_count or 0
-        self.set_rmax(rmax=self.rmax+calm_count)
+        self.set_rmax(rmax=self.rmax + calm_count)
         self.set_radii_angle(angle=self.radii_angle)
 
     def legend(self, loc='lower left', **kwargs):
@@ -179,7 +179,7 @@ class WindroseAxes(PolarAxes):
                 else:
                     raise AttributeError("Can't handle patches")
                 handles.append(mpl.patches.Rectangle((0, 0), 0.2, 0.2,
-                               facecolor=color, edgecolor='black'))
+                                                     facecolor=color, edgecolor='black'))
             return handles
 
         def get_labels(decimal_places=1):
@@ -299,12 +299,13 @@ class WindroseAxes(PolarAxes):
             mask = var > calm_limit
             self.calm_count = len(var) - np.count_nonzero(mask)
             if normed:
-                self.calm_count = self.calm_count*100/len(var)
+                self.calm_count = self.calm_count * 100 / len(var)
             var = var[mask]
             direction = direction[mask]
 
         # Set the global information dictionnary
-        self._info['dir'], self._info['bins'], self._info['table'] = histogram(direction, var, bins, nsector, normed, blowto)
+        self._info['dir'], self._info['bins'], self._info['table'] = histogram(direction, var, bins, nsector, normed,
+                                                                               blowto)
 
         return bins, nbins, nsector, colors, angles, kwargs
 
@@ -315,9 +316,9 @@ class WindroseAxes(PolarAxes):
         """
         if self.calm_count and self.calm_count > 0:
             circle = mpl.patches.Circle((0., 0.), self.calm_count,
-                            transform=self.transData._b,
-                            color=CALM_CIRCLE_COLOR,
-                            alpha=CALM_CIRCLE_ALPHA)
+                                        transform=self.transData._b,
+                                        color=CALM_CIRCLE_COLOR,
+                                        alpha=CALM_CIRCLE_ALPHA)
             self.add_artist(circle)
         return self.calm_count or 0
 
@@ -348,7 +349,7 @@ class WindroseAxes(PolarAxes):
         in different colors in the order specified.
         * cmap : a cm Colormap instance from matplotlib.cm.
           - if cmap == None and colors == None, a default Colormap is used.
-        * calm_limit: float - Calm limit for the var parameter. If not None, 
+        * calm_limit: float - Calm limit for the var parameter. If not None,
         a centered red circle will be draw for representing the calms occurences
         and all datas below this value will be removed from the computation
 
@@ -361,8 +362,8 @@ class WindroseAxes(PolarAxes):
         # closing lines
         angles = np.hstack((angles, angles[-1] - 2 * np.pi / nsector))
         vals = np.hstack((self._info['table'],
-                         np.reshape(self._info['table'][:, 0],
[…]
