### 2017-08-02 python-windrose/windrose#65 @scls19fr
Ok I will do
I will probably need to retag
