### 2017-08-02 a55addf
Closes #65
