### 2017-08-02 python-windrose/windrose#65 @scls19fr
Fixed but according both https://spdx.org/licenses/ and https://opensource.org/licenses/alphabetical CeCILL-B is not listed in OSI approved list of license (despite the fact that http://www.cecill.info/licences.en.html claims that it has the caracteristics of BSD-like licenses)
I just hope it won't be an issue for Journal of Open Source Software review.
