+++ windrose/version.py
@@ -1,7 +1,7 @@
 __author__ = 'Lionel Roubeyrie'
 __credits__ = ['Sebastien Celles']
 __email__ = 's.celles@gmail.com'
-__version__ = '1.6.1'
+__version__ = '1.6.2'
 __license__ = 'CeCILL-B'
 __url__ = 'https://github.com/python-windrose/windrose'
 __maintainer__ = "Sebastien Celles"
