### 2017-08-02 python-windrose/windrose#65 @scls19fr
Reopening this issue because of pip upload fail!

```
Submitting dist/windrose-1.6.2.tar.gz to https://upload.pypi.org/legacy/
Upload failed (400): classifiers: 'License :: CEA CNRS Inria Logiciel Libre License B (CeCILL-B)' is not a valid choice for this field
error: Upload failed (400): classifiers: 'License :: CEA CNRS Inria Logiciel Libre License B (CeCILL-B)' is not a valid choice for this field```
