### 2017-08-02 python-windrose/windrose#65 @scls19fr
I also tried

```
    classifiers=[
        # Pick your license as you wish (should match "license" above)
        'CeCILL-B',
    ],
```

and

```
    classifiers=[
        'License :: CeCILL-B',
    ],
```

and get similar error

`Upload failed (400): classifiers: 'License :: CeCILL-B' is not a valid choice for this field`
