### 2017-08-02 python-windrose/windrose#65 @scls19fr
I've just opened an issue on PyPi https://github.com/pypa/pypi-legacy/issues/679 to ask to add CeCILL-B to allowed classifiers in setup.py or to provide a better workaround.
