### 2017-08-02 072bef1
Comment License field in classifier
