### 2017-08-03 python-windrose/windrose#65 @LionelR
@scls19fr for consistency with the primary license and to be simple in the process, I think we can use the MIT license as a substitute.
