### 2017-08-03 python-windrose/windrose#65 @scls19fr
I pushed code to PyPi without defining License in classifier (as suggested in https://github.com/pypa/pypi-legacy/issues/679 )
License is defined to CeCILL-B in this release.
If that's an issue for JOSS I will create an other release with dual license.
