### 2017-08-03 #268 @scls19fr
PyPi doesn't support CeCILL-B license in classifiers for now (see https://github.com/pypa/pypi-legacy/issues/679 ) so as a workaround I commented out License field in classifier but licence is (as previously) correctly set to CeCILL-B thanks to https://github.com/python-windrose/windrose/blob/master/setup.py#L35 and https://github.com/python-windrose/windrose/blob/master/windrose/version.py#L5 )
I can be seen on https://pypi.python.org/pypi?:action=display&name=windrose&version=1.6.2
