### 2017-08-15 #268 @arfon
Hi @scls19fr, quoting from https://github.com/python-windrose/windrose/issues/65#issuecomment-320047090:

> If that's an issue for JOSS I will create an other release with dual license.

If possible, I think it might be easier to dual-license this project. We're not in a position to offer legal advice to authors but I would be more comfortable if there was a license file present in the repository that was definitively listed in the OSI approved licenses.
