### 2017-08-16 python-windrose/windrose#69 @scls19fr
Dual licensing
According to https://github.com/openjournals/joss-reviews/issues/268#issuecomment-322431025
we need dual licensing (OSI approved licenses).

@LionelR what OSI approved license should we use?

https://pypi.python.org/pypi?%3Aaction=list_classifiers

    License :: OSI Approved :: MIT License

(like you suggested in https://github.com/python-windrose/windrose/issues/65#issuecomment-319883719 )

or

    License :: OSI Approved :: BSD License

because according http://www.cecill.info/licences.en.html
CeCILL-B: "B" because it has the caracteristics of BSD-like licenses

This could help
https://opensource.stackexchange.com/questions/217/what-are-the-essential-differences-between-bsd-and-mit-licences
