### 2017-08-16 python-windrose/windrose#69 @LionelR
Ok for BSD license

Le 16 août 2017 13:10:47 GMT+02:00, scls19fr <notifications@github.com> a écrit :
>According to
>https://github.com/openjournals/joss-reviews/issues/268#issuecomment-322431025
>we need dual licensing (OSI approved licenses).
>
>@LionelR what OSI approved license should we use?
>
>https://pypi.python.org/pypi?%3Aaction=list_classifiers
>
>    License :: OSI Approved :: MIT License
>
>(like you suggested in
>https://github.com/python-windrose/windrose/issues/65#issuecomment-319883719
>)
>
>
>or
>
>    License :: OSI Approved :: BSD License
>
>because according http://www.cecill.info/licences.en.html
>CeCILL-B: "B" because it has the caracteristics of BSD-like licenses
>
>This could help
>https://opensource.stackexchange.com/questions/217/what-are-the-essential-differences-between-bsd-and-mit-licences
>
>--
>You are receiving this because you were mentioned.
>Reply to this email directly or view it on GitHub:
>https://github.com/python-windrose/windrose/issues/69

--
Envoyé de mon appareil Android avec Courriel K-9 Mail. Veuillez excuser ma brièveté.
