### 2017-08-16 python-windrose/windrose#69 @scls19fr
2-clause BSD License (BSD-2-Clause)

or

3-clause BSD License (BSD-3-Clause)

are OSI approved not 4-clause BSD licence

So now we have to choose between BSD-2-Clause and BSD-3-Clause
