### 2017-08-22 7373bff
Dual licensing Closes #69 Add a release procedure Closes #59
