### 2017-08-22 #268 @scls19fr
windrose is now published with dual licensing: `CeCILL-B OR BSD-3-Clause`
