### 2017-09-04 python-windrose/windrose#69 @scls19fr
According https://github.com/pypa/pypi-legacy/issues/679#issuecomment-326993025

    License :: CeCILL-B Free Software License Agreement (CECILL-B)

is now a possible classifier

Instead of https://github.com/python-windrose/windrose/blob/ec7b1ed9fd9a9e28638eb0c21928ace4e9a4f39f/setup.py#L67
