### 2018-01-19 d77f49c
facecolor not recognized here
