### 2018-03-15 #268 @amoeba
Hey @arfon yes I would. I can tackle it this weekend.
