### 2018-03-16 #268 @arfon
@scls19fr - please could you take a look at the guidelines for the [JOSS paper structure](http://joss.theoj.org/about#paper_structure) and add some more detail to your `paper.md` please?
