### 2018-03-19 python-windrose/windrose#77 @amoeba
JOSS Review
Hey @scls19fr I've taken a look at your submission, https://github.com/openjournals/joss-reviews/issues/268,
 and found most everything to be in order, with some minor comments. Overall, I appreciate the job you've done making installation easy and getting started very straightforward. I don't analyze wind data often but I think I would able to use Windrose pretty easily if I ever did.

I've re-visited the reviewer checklist, which was left in an incomplete state due to the last reviewer, and verified every checked item and either checked or unchecked the ones I had comments on. Below are my comments. I'd like to see you at least address my comments on versions and documentation prior to my acceptance.

> License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

I checked this because I saw the above conversation with @arfon and assumed that dual licensing was both okay and done in a suitable fashion.

> Version: Does the release version given match the GitHub release (1.6)?

No, version latest release, which addresses the previous licensing issues, is 1.6.3. 1.6.3 is also the version installed via pip and the version I tested the functionality of. We should update the JOSS submission and archive so its for the 1.6.3 release.

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g. API method documentation)?

The documentation that is present is pretty good (especially the readme) but I think it could be more comprehensive, especially where completely absent and where `**kwargs` is used for parameters. The most complete documentation for this package was in the docstrings which I had to browse the source to read. I think a dedicated and more complete documentation site (such as those we often generate with readthedocs) would be super helpful for new and existing users. What do you think about getting something like this working, @scls19fr? You might consider trimming the readme a bit and putting the content contained there the aforementioned documentation. Alterantively, you may find it easier to just write more documentation into the readme, which would be fine.

> Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

There are manual examples which may serve as tests and a basic set of automated functionality tests included with the package, though there were no unit or regression tests (which could be nice!). Since this is a plotting package, a large amount of the functionality, and therefore potential places for bugs in the user experience depends on matplotlib, I wonder what the author would think about including automated or semi-automated visual regression tests. An easy first pass at this could just involve setting a random seed at the top of the visual regression test suite, adding some tests which create the various kinds of plots the package produces, and diffing the output with version-controlled copies from the last test run.

> Software paper

I'll wait to comment on this until the author updates the paper as per the editor's notes.

> Other notes:

These are general comments and don't need to be addressed for the JOSS submission:

- Noticed some markdown headers in windrose_sample_random.ipynb are missing a space between the # and the header which prevents proper formatting
- Noticed the global, `RESOLUTION`, is not actually used. Might be good to lint the source to find similar issues.
- When I ran this code:

    ```py
    ax = WindroseAxes.from_ax()
    ax.contourf(wd, ws, bins=np.arange(0, 8, 1), cmap=cm.hot)
    ax.set_legend()
    ```

    I got this warning:

    > /Users/bryce/src/windrose/windrose/windrose.py:394: MatplotlibDeprecationWarning: The poly_between function was deprecated in version 2.2.
      xs, ys = poly_between(angles, 0, val)
