### 2018-03-19 #268 @amoeba
Hey @scls19fr I've posted my partial review at https://github.com/python-windrose/windrose/issues/77 while we wait on the updates you are making to the paper. Let me know if you have any questions over there.
