### 2018-03-19 python-windrose/windrose#77 @scls19fr
Thanks @amoeba for your review.
I will tackle issues you mentionned during the next weeks (I will be quite busy the 2 following weeks because of a high teaching volume) and will be back here.
Thanks also @arfon for reviving https://github.com/openjournals/joss-reviews/issues/268
