+++ docs/development.rst
@@ -1,0 +1,54 @@
+Development
+-----------
+
+You can help to develop this library.
+
+Issues
+~~~~~~
+
+You can submit issues using
+https://github.com/python-windrose/windrose/issues
+
+Clone
+~~~~~
+
+You can clone repository to try to fix issues yourself using:
+
+.. code:: bash
+
+    $ git clone https://github.com/python-windrose/windrose.git
+
+Run unit tests
+~~~~~~~~~~~~~~
+
+Run all unit tests
+
+.. code:: bash
+
+    $ nosetests -s -v
+
+Run a given test
+
+.. code:: bash
+
+    $ nosetests tests.test_windrose:test_plot_by -s -v
+
+Install development version
+~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. code:: bash
+
+    $ python setup.py install
+
+or
+
+.. code:: bash
+
+    $ sudo pip install git+https://github.com/python-windrose/windrose.git
+
+Collaborating
+~~~~~~~~~~~~~
+
+-  Fork repository
+-  Create a branch which fix a given issue
+-  Submit pull requests
