+++ windrose/windrose.py
@@ -12,7 +12,6 @@ from numpy.lib.twodim_base import histogram2d
 import matplotlib.pyplot as plt
 from pylab import poly_between
 
-RESOLUTION = 100
 ZBASE = -1000  # The starting zorder for all drawing, negative to have the grid on
 VAR_DEFAULT = 'speed'
 DIR_DEFAULT = 'direction'
