### 2018-05-01 python-windrose/windrose#77 @scls19fr
@amoeba Sphynx doc is now online and available at http://windrose.readthedocs.io/en/latest/

I'm using `flake8` as linter and Travis is running after each commit, but it didn't catch the problem you mentioned about `RESOLUTION` constant. Anyway, I removed this unused constant.

About the depreciation warning
`MatplotlibDeprecationWarning: The poly_between function was deprecated in version 2.2.`
should be fixed in #81
https://www.google.fr/search?q=MatplotlibDeprecationWarning%3A+The+poly_between+function+was+deprecated+in+version+2.2. doesn't give much help about a workaround.

windrose_sample_random.ipynb markdown rendering have been fixed

About tests, there is a https://github.com/python-windrose/windrose/blob/master/tests/test_windrose.py visual regression can be catched manually with generated images in test/output directory, but not automatically.
These tests although insure automatically that code doesn't break. Automated visual regression tests is out of my knowledge.
