### 2018-05-01 #268 @scls19fr
I've just reply today to @amoeba in https://github.com/python-windrose/windrose/issues/77#issuecomment-385725732 to address most of his comments
Sorry for delay.
