### 2018-05-01 python-windrose/windrose#77 @amoeba
Thanks @scls19fr I'll take a look this week.
