### 2018-05-02 f928538
Update usage.rst
Fix bullet list
