+++ docs/usage.rst
@@ -142,32 +142,40 @@ windrose, calling help(ax.bar) will give :
 ``windrose.WindroseAxes`` instance Plot a windrose in bar mode. For each
 var bins and for each sector, a colored bar will be draw on the axes.
 
-Mandatory: - ``direction`` : 1D array - directions the wind blows from,
-North centred - ``var`` : 1D array - values of the variable to compute.
-Typically the wind speeds
+Mandatory:
 
-Optional: - ``nsector`` : integer - number of sectors used to compute
+- ``direction`` : 1D array
+- directions the wind blows from, North centred
+- ``var`` : 1D array
+- values of the variable to compute. Typically the wind speeds
+
+Optional: 
+
+- ``nsector`` : integer - number of sectors used to compute
 the windrose table. If not set, nsectors=16, then each sector will be
 360/16=22.5°, and the resulting computed table will be aligned with the
-cardinals points. - ``bins`` : 1D array or integer- number of bins, or a
+cardinals points. 
+- ``bins`` : 1D array or integer- number of bins, or a
 sequence of bins variable. If not set, bins=6 between min(var) and
-max(var). - ``blowto`` : bool. If True, the windrose will be pi rotated,
-to show where the wind blow to (usefull for pollutant rose). -
+max(var). 
+- ``blowto`` : bool. If True, the windrose will be pi rotated,
+to show where the wind blow to (usefull for pollutant rose). 
+-
 ``colors`` : string or tuple - one string color (``'k'`` or
 ``'black'``), in this case all bins will be plotted in this color; a
 tuple of matplotlib color args (string, float, rgb, etc), different
-levels will be plotted in different colors in the order specified. -
-``cmap`` : a cm Colormap instance from ``matplotlib.cm``. - if
+levels will be plotted in different colors in the order specified. 
+- ``cmap`` : a cm Colormap instance from ``matplotlib.cm``. - if
 ``cmap == None`` and ``colors == None``, a default Colormap is used. -
 ``edgecolor`` : string - The string color each edge bar will be plotted.
 Default : no edgecolor - ``opening`` : float - between 0.0 and 1.0, to
-control the space between each sector (1.0 for no space) -
-``mean_values`` : Bool - specify wind speed statistics with
+control the space between each sector (1.0 for no space) 
+- ``mean_values`` : Bool - specify wind speed statistics with
 direction=specific mean wind speeds. If this flag is specified, var is
 expected to be an array of mean wind speeds corresponding to each entry
 in ``direction``. These are used to generate a distribution of wind
-speeds assuming the distribution is Weibull with shape factor = 2. -
-``weibull_factors`` : Bool - specify wind speed statistics with
+speeds assuming the distribution is Weibull with shape factor = 2. 
+- ``weibull_factors`` : Bool - specify wind speed statistics with
 direction=specific weibull scale and shape factors. If this flag is
 specified, var is expected to be of the form [[7,2], ...., [7.5,1.9]]
 where var[i][0] is the weibull scale factor and var[i][1] is the shape
@@ -225,11 +233,15 @@ previously shown.
 Mandatory: - ``df``: Pandas DataFrame with ``DateTimeIndex`` as index
 and at least 2 columns (``'speed'`` and ``'direction'``).
 
-Optional: - ``kind`` : kind of plot (might be either, ``'contour'``,
-``'contourf'``, ``'bar'``, ``'box'``, ``'pdf'``) - ``var_name`` : name
+Optional: 
+
+- ``kind`` : kind of plot (might be either, ``'contour'``,
+``'contourf'``, ``'bar'``, ``'box'``, ``'pdf'``) 
+- ``var_name`` : name
[…]
