### 2018-05-02 60f5549
Update usage.rst
Fix other bullet lists and removed some text
