### 2018-05-02 45c9ae0
Update usage.rst
Fix an other bullet list
