### 2018-05-02 e66efb6
Update usage.rst
Continuing text must be aligned in bullet lists
