### 2018-05-02 6add617
Update usage.rst
