### 2018-05-02 25d53a7
Improve docstrings rendering
Closes #82
