+++ windrose/windrose.py
@@ -20,9 +20,26 @@ DPI_DEFAULT = 80
 
 
 class WindAxesFactory(object):
-    # Create based on class name:
+    """
+
+    Factory class to create WindroseAxes or WindAxes
+
+    """
     @staticmethod
     def create(typ, ax=None, *args, **kwargs):
+        """
+
+        Create 
+
+        Mandatory:
+
+        * typ 'windroseaxes' to create WindroseAxes or 'windaxes' to create WindAxes
+
+        Optional:
+
+        * ax Matplotlib axe
+
+        """
         typ = typ.lower()
         d = {
             'windroseaxes': WindroseAxes,
@@ -304,23 +321,23 @@ class WindroseAxes(PolarAxes):
 
         * direction : 1D array - directions the wind blows from, North centred
         * var : 1D array - values of the variable to compute. Typically the wind
-        speeds
+          speeds
 
         Optional:
 
         * nsector: integer - number of sectors used to compute the windrose
-        table. If not set, nsectors=16, then each sector will be 360/16=22.5°,
-        and the resulting computed table will be aligned with the cardinals
-        points.
+          table. If not set, nsectors=16, then each sector will be 360/16=22.5°,
+          and the resulting computed table will be aligned with the cardinals
+          points.
         * bins : 1D array or integer- number of bins, or a sequence of
-        bins variable. If not set, bins=6, then
-            bins=linspace(min(var), max(var), 6)
+          bins variable. If not set, bins=6, then
+          bins=linspace(min(var), max(var), 6)
         * blowto : bool. If True, the windrose will be pi rotated,
-        to show where the wind blow to (usefull for pollutant rose).
+          to show where the wind blow to (usefull for pollutant rose).
         * colors : string or tuple - one string color ('k' or 'black'), in this
-        case all bins will be plotted in this color; a tuple of matplotlib
-        color args (string, float, rgb, etc), different levels will be plotted
-        in different colors in the order specified.
+          case all bins will be plotted in this color; a tuple of matplotlib
+          color args (string, float, rgb, etc), different levels will be plotted
+          in different colors in the order specified.
         * cmap : a cm Colormap instance from matplotlib.cm.
           - if cmap == None and colors == None, a default Colormap is used.
 
@@ -357,23 +374,23 @@ class WindroseAxes(PolarAxes):
 
         * direction : 1D array - directions the wind blows from, North centred
         * var : 1D array - values of the variable to compute. Typically the wind
-        speeds
+          speeds
 
         Optional:
 
         * nsector: integer - number of sectors used to compute the windrose
-        table. If not set, nsectors=16, then each sector will be 360/16=22.5°,
-        and the resulting computed table will be aligned with the cardinals
-        points.
+          table. If not set, nsectors=16, then each sector will be 360/16=22.5°,
+          and the resulting computed table will be aligned with the cardinals
+          points.
         * bins : 1D array or integer- number of bins, or a sequence of
-        bins variable. If not set, bins=6, then
-            bins=linspace(min(var), max(var), 6)
+          bins variable. If not set, bins=6, then
+          bins=linspace(min(var), max(var), 6)
         * blowto : bool. If True, the windrose will be pi rotated,
-        to show where the wind blow to (usefull for pollutant rose).
+          to show where the wind blow to (usefull for pollutant rose).
         * colors : string or tuple - one string color ('k' or 'black'), in this
-        case all bins will be plotted in this color; a tuple of matplotlib
-        color args (string, float, rgb, etc), different levels will be plotted
-        in different colors in the order specified.
+          case all bins will be plotted in this color; a tuple of matplotlib
+          color args (string, float, rgb, etc), different levels will be plotted
+          in different colors in the order specified.
         * cmap : a cm Colormap instance from matplotlib.cm.
[…]
