### 2018-05-04 9b6f592
correct docstring to comply with reStructuredText
