+++ windrose/windrose.py
@@ -19,10 +19,11 @@ DIR_DEFAULT = 'direction'
 FIGSIZE_DEFAULT = (8, 8)
 DPI_DEFAULT = 80
 
+
 def _autogen_docstring(base):
     """Autogenerated wrappers will get their docstring from a base function
     with an addendum."""
-    #msg = "\n\nAdditional kwargs: hold = [True|False] overrides default hold state"
+    # msg = "\n\nAdditional kwargs: hold = [True|False] overrides default hold state"
     msg = ''
     addendum = docstring.Appender(msg, '\n\n')
     return lambda func: addendum(docstring.copy_dedent(base)(func))
@@ -38,15 +39,19 @@ class WindAxesFactory(object):
     def create(typ, ax=None, *args, **kwargs):
         """
 
-        Create 
+        Create
 
         Mandatory:
 
-        * typ 'windroseaxes' to create WindroseAxes or 'windaxes' to create WindAxes
+        Parameters
+        ----------
+        typ : string, 'windroseaxes' or 'windaxes'
+            Type of axes to create
+                * windroseaxes : a WindroseAxes axe
+                * windaxe : a WindAxes axe
 
-        Optional:
-
-        * ax Matplotlib axe
+        ax : matplotlib.Axes, optional
+            A matplotlib axe
 
         """
         typ = typ.lower()
@@ -165,7 +170,7 @@ class WindroseAxes(PolarAxes):
 
         loc : int, string or pair of floats, default: 'lower left'
             see :matplotlib.pyplot.legend:
-        
+
         decimal_places : int, default 1
             The decimal places of the formated legend
 
@@ -345,7 +350,7 @@ class WindroseAxes(PolarAxes):
             float, rgb, etc), different levels will be plotted in different
             colors in the order specified.
         cmap : a cm Colormap instance from matplotlib.cm, optional
-  	        if cmap == None and colors == None, a default Colormap is used.
+            if cmap == None and colors == None, a default Colormap is used.
 
         others kwargs : see help(pylab.plot)
 
@@ -379,7 +384,7 @@ class WindroseAxes(PolarAxes):
         Parameters
         ----------
         direction : 1D array
-      	    directions the wind blows from, North centred
+            directions the wind blows from, North centred
         var : 1D array
             values of the variable to compute. Typically the wind speeds
 
@@ -399,7 +404,7 @@ class WindroseAxes(PolarAxes):
             float, rgb, etc), different levels will be plotted in different
             colors in the order specified.
         cmap : a cm Colormap instance from matplotlib.cm, optional
-	          if cmap == None and colors == None, a default Colormap is used.
+            if cmap == None and colors == None, a default Colormap is used.
 
         others kwargs : see help(pylab.plot)
 
@@ -503,7 +508,7 @@ class WindroseAxes(PolarAxes):
             directions the wind blows from, North centred
         var : 1D array
             values of the variable to compute. Typically the wind speeds
-       
+
         nsector: integer, optional
             number of sectors used to compute the windrose table. If not set,
             nsectors=16, then each sector will be 360/16=22.5°, and the
@@ -598,16 +603,16 @@ def histogram(direction, var, bins, nsector, normed=False, blowto=False):
     Parameters
     ----------
     direction : 1D array
-	      directions the wind blows from, North centred
+        directions the wind blows from, North centred
     var : 1D array
         values of the variable to compute. Typically the wind speeds
     bins : list
-	      list of var category against we're going to compute the table
+        list of var category against we're going to compute the table
     nsector : integer
-	      number of sectors
+        number of sectors
 
[…]
