### 2018-05-05 f417de8
[DOC] comply docstring with numpydoc (#83)
* correct docstring to comply with reStructuredText

* minor fix (trailing space)

* PEP8
