+++ windrose/windrose.py
@@ -5,6 +5,7 @@
 
 import locale
 import matplotlib as mpl
+from matplotlib import docstring
 import numpy as np
 import random
 from matplotlib.projections.polar import PolarAxes
@@ -19,6 +20,15 @@ FIGSIZE_DEFAULT = (8, 8)
 DPI_DEFAULT = 80
 
 
+def _autogen_docstring(base):
+    """Autogenerated wrappers will get their docstring from a base function
+    with an addendum."""
+    # msg = "\n\nAdditional kwargs: hold = [True|False] overrides default hold state"
+    msg = ''
+    addendum = docstring.Appender(msg, '\n\n')
+    return lambda func: addendum(docstring.copy_dedent(base)(func))
+
+
 class WindAxesFactory(object):
     """
 
@@ -29,15 +39,19 @@ class WindAxesFactory(object):
     def create(typ, ax=None, *args, **kwargs):
         """
 
-        Create 
+        Create
 
         Mandatory:
 
-        * typ 'windroseaxes' to create WindroseAxes or 'windaxes' to create WindAxes
+        Parameters
+        ----------
+        typ : string, 'windroseaxes' or 'windaxes'
+            Type of axes to create
+                * windroseaxes : a WindroseAxes axe
+                * windaxe : a WindAxes axe
 
-        Optional:
-
-        * ax Matplotlib axe
+        ax : matplotlib.Axes, optional
+            A matplotlib axe
 
         """
         typ = typ.lower()
@@ -80,6 +94,9 @@ class WindroseAxes(PolarAxes):
 
     @staticmethod
     def from_ax(ax=None, fig=None, rmax=None, *args, **kwargs):
+        """
+        Return a WindroseAxes object for the figure `fig`.
+        """
         if ax is None:
             if fig is None:
                 fig = plt.figure(figsize=FIGSIZE_DEFAULT, dpi=DPI_DEFAULT, facecolor='w', edgecolor='w')
@@ -144,40 +161,29 @@ class WindroseAxes(PolarAxes):
             self.set_rmax(rmax=self.rmax)
         self.set_radii_angle(angle=self.radii_angle)
 
-    def legend(self, loc='lower left', **kwargs):
+    def legend(self, loc='lower left', decimal_places=1, **kwargs):
         """
         Sets the legend location and her properties.
-        The location codes are
 
-          'best'         : 0,
-          'upper right'  : 1,
-          'upper left'   : 2,
-          'lower left'   : 3,
-          'lower right'  : 4,
-          'right'        : 5,
-          'center left'  : 6,
-          'center right' : 7,
-          'lower center' : 8,
-          'upper center' : 9,
-          'center'       : 10,
+        Parameters
+        ----------
 
-        If none of these are suitable, loc can be a 2-tuple giving x,y
-        in axes coords, ie,
+        loc : int, string or pair of floats, default: 'lower left'
+            see :matplotlib.pyplot.legend:
 
-          loc = (0, 1) is left top
-          loc = (0.5, 0.5) is center, center
+        decimal_places : int, default 1
+            The decimal places of the formated legend
 
-        and so on.  The following kwargs are supported:
+        The following kwargs are supported:
 
-        isaxes=True           # whether this is an axes legend
-        prop = FontProperties(size='smaller')  # the font property
-        pad = 0.2             # the fractional whitespace inside the legend border
-        shadow                # if True, draw a shadow behind legend
-        labelsep = 0.005     # the vertical space between the legend entries
-        handlelen = 0.05     # the length of the legend lines
-        handletextsep = 0.02 # the space between the legend line and legend text
-        borderaxespad = 0.02       # the border between the axes and legend edge
-        decimal_places = 1   # the decimal places of the formated legend
[…]
