### 2018-05-05 3f0a335
[DOC] comply the legend doc to numpydoc format
