### 2018-05-05 75bd97e
[DOC] support interpshinx and numpydoc style for sphinx
