+++ docs/conf.py
@@ -31,7 +31,11 @@ sys.path.insert(0, os.path.abspath('../windrose'))
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.autodoc']
+extensions = [
+    'sphinx.ext.autodoc',
+    'sphinx.ext.intersphinx',
+    'sphinxcontrib.napoleon',
+]
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -169,4 +173,6 @@ texinfo_documents = [
 ]
 
 
-
+interpshinx_mapping = {
+    "matplotlib": ('http://matplotlib.org', None),
+}
