### 2018-05-05 febd8a4
[DOC] update docstring
