+++ windrose/windrose.py
@@ -245,7 +245,7 @@ class WindroseAxes(PolarAxes):
         plt.setp(legend.get_texts(), fontsize=8)
         return legend
 
-    def _init_plot(self, direction, var, **kwargs):
+    def _init_plot(self, direction, var, normed=False, blowto=False, **kwargs):
         """
         Internal method used by all plotting commands
 
@@ -255,6 +255,16 @@ class WindroseAxes(PolarAxes):
             directions the wind blows from, North centred
         var : 1D array,
             values of the variable to compute. Typically the wind speeds
+
+        Other Parameters
+        ----------------
+        normed : boolean, default False
+        blowto : boolean, default False
+        weibull_factors :
+        mean_values :
+        frequency :
+        kwarg
+            Any argument accepted by :obj:`matplotlib.pyplot.plot`.
         """
 
         # if weibull factors are entered overwrite direction and var
@@ -322,8 +332,8 @@ class WindroseAxes(PolarAxes):
         # Building the angles list
         angles = np.arange(0, -2 * np.pi, -2 * np.pi / nsector) + np.pi / 2
 
-        normed = kwargs.pop('normed', False)
-        blowto = kwargs.pop('blowto', False)
+        # normed = kwargs.pop('normed', False)
+        # blowto = kwargs.pop('blowto', False)
 
         # Set the global information dictionnary
         self._info['dir'], self._info['bins'], self._info['table'] = histogram(direction, var, bins, nsector, normed, blowto)
@@ -344,6 +354,8 @@ class WindroseAxes(PolarAxes):
         var : 1D array
             values of the variable to compute. Typically the wind speeds.
 
+        Other Parameters
+        ----------------
         sector : integer, optional
             number of sectors used to compute the windrose table. If not set,
             nsectors=16, then each sector will be 360/16=22.5°, and the
@@ -359,10 +371,10 @@ class WindroseAxes(PolarAxes):
             plotted in this color; a tuple of matplotlib color args (string,
             float, rgb, etc), different levels will be plotted in different
             colors in the order specified.
-        cmap : a cm Colormap instance from matplotlib.cm, optional
+        cmap : a cm Colormap instance from :obj:`matplotlib.cm`, optional
             if cmap == None and colors == None, a default Colormap is used.
-
-        others kwargs : see help(pylab.plot)
+        others kwargs
+            Any supported argument of :obj:`matplotlib.pyplot.plot`
 
         """
         bins, nbins, nsector, colors, angles, kwargs = self._init_plot(direction, var,
@@ -398,6 +410,8 @@ class WindroseAxes(PolarAxes):
         var : 1D array
             values of the variable to compute. Typically the wind speeds
 
+        Other Parameters
+        ----------------
         nsector: integer, optional
             number of sectors used to compute the windrose table. If not set,
             nsectors=16, then each sector will be 360/16=22.5°, and the
@@ -413,11 +427,11 @@ class WindroseAxes(PolarAxes):
             plotted in this color; a tuple of matplotlib color args (string,
             float, rgb, etc), different levels will be plotted in different
             colors in the order specified.
-        cmap : a cm Colormap instance from matplotlib.cm, optional
+        cmap : a cm Colormap instance from :obj:`matplotlib.cm`, optional
             if cmap == None and colors == None, a default Colormap is used.
 
-        others kwargs : see help(pylab.plot)
-
+        others kwargs
+            Any supported argument of :obj:`matplotlib.pyplot.plot`
         """
 
         bins, nbins, nsector, colors, angles, kwargs = self._init_plot(direction, var,
@@ -452,6 +466,8 @@ class WindroseAxes(PolarAxes):
         var : 1D array
             values of the variable to compute. Typically the wind speeds.
 
+        Other Parameters
+        ----------------
         nsector : integer, optional
[…]
