### 2018-05-05 d0be1d8
hotfix: wring extension in conf.py
