+++ docs/conf.py
@@ -34,7 +34,7 @@ sys.path.insert(0, os.path.abspath('../windrose'))
 extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.intersphinx',
-    'sphinxcontrib.napoleon',
+    'sphinx.ext.napoleon',
 ]
 
 # Add any paths that contain templates here, relative to this directory.
