### 2018-05-05 b473f0e
hotfix: revert unwanted change + update docstring
