### 2018-05-05 ee726e4
[DOC] improve documentation (#84)
* [DOC] comply the legend doc to numpydoc format

* [DOC] support interpshinx and numpydoc style for sphinx

* [DOC] update docstring

* hotfix: wring extension in conf.py

* hotfix: revert unwanted change + update docstring
