+++ windrose/windrose.py
@@ -167,23 +167,33 @@ class WindroseAxes(PolarAxes):
 
         Parameters
         ----------
-
         loc : int, string or pair of floats, default: 'lower left'
-            see :matplotlib.pyplot.legend:
+            see :obj:`matplotlib.pyplot.legend`.
 
         decimal_places : int, default 1
             The decimal places of the formated legend
 
-        The following kwargs are supported:
-
-        * isaxes=True           # whether this is an axes legend
-        * prop = FontProperties(size='smaller')  # the font property
-        * pad = 0.2             # the fractional whitespace inside the legend border
-        * shadow                # if True, draw a shadow behind legend
-        * labelsep = 0.005     # the vertical space between the legend entries
-        * handlelen = 0.05     # the length of the legend lines
-        * handletextsep = 0.02 # the space between the legend line and legend text
-        * borderaxespad = 0.02       # the border between the axes and legend edge
+        Other Parameters
+        ----------------
+        isaxes : boolean, default True
+            whether this is an axes legend
+        prop : FontProperties(size='smaller')
+            the font property
+        borderpad : float
+            the fractional whitespace inside the legend border
+        shadow : boolean
+            if True, draw a shadow behind legend
+        labelspacing : float, 0.005
+            the vertical space between the legend entries
+        handlelenght : float, 0.05
+            the length of the legend lines
+        handletextsep : float, 0.02
+            the space between the legend line and legend text
+        borderaxespad : float, 0.02
+            the border between the axes and legend edge
+        kwarg
+            Every other kwarg argument supported by
+            :obj:`matplotlib.pyplot.legend`
         """
 
         def get_handles():
@@ -245,6 +255,21 @@ class WindroseAxes(PolarAxes):
             directions the wind blows from, North centred
         var : 1D array,
             values of the variable to compute. Typically the wind speeds
+
+        Other Parameters
+        ----------------
+        normed : boolean, default False
+        blowto : boolean, default False
+        colors : str or list of str, default None
+            The colors of the plot.
+        cmap : color map, default `jet`
+            A :obj:`matplotlib.cm` colormap for the plot.
+            Warning! It overrides `colors`.
+        weibull_factors :
+        mean_values :
+        frequency :
+        kwarg
+            Any argument accepted by :obj:`matplotlib.pyplot.plot`.
         """
 
         # if weibull factors are entered overwrite direction and var
@@ -334,6 +359,8 @@ class WindroseAxes(PolarAxes):
         var : 1D array
             values of the variable to compute. Typically the wind speeds.
 
+        Other Parameters
+        ----------------
         sector : integer, optional
             number of sectors used to compute the windrose table. If not set,
             nsectors=16, then each sector will be 360/16=22.5°, and the
@@ -349,10 +376,10 @@ class WindroseAxes(PolarAxes):
             plotted in this color; a tuple of matplotlib color args (string,
             float, rgb, etc), different levels will be plotted in different
             colors in the order specified.
-        cmap : a cm Colormap instance from matplotlib.cm, optional
+        cmap : a cm Colormap instance from :obj:`matplotlib.cm`, optional
             if cmap == None and colors == None, a default Colormap is used.
-
-        others kwargs : see help(pylab.plot)
+        others kwargs
+            Any supported argument of :obj:`matplotlib.pyplot.plot`
 
         """
         bins, nbins, nsector, colors, angles, kwargs = self._init_plot(direction, var,
@@ -388,6 +415,8 @@ class WindroseAxes(PolarAxes):
         var : 1D array
             values of the variable to compute. Typically the wind speeds
 
+        Other Parameters
+        ----------------
         nsector: integer, optional
[…]
