### 2018-05-07 python-windrose/windrose#77 @amoeba
Hi @scls19fr thanks for all the work to address my comments. Things are looking really good. I have two outstanding reviewer checklist items.

### Version

The JOSS version is still at 1.6, the latest GitHub release is 1.6.3, and I suspect you'll want JOSS to archive yet another version that includes all of your good changes. I'm not sure if I can help with this.

### JOSS Paper

- The paper.md could use a _little_ more work. @arfon had asked for a bit more detail in the body of the paper and I think that's a great suggestion. Preferably at least a solid paragraph but potentially a few. Previous JOSS submissions would be a good place to look for ideas on style.
- Are there any references you could make to past work? I know this project builds on previous work so it'd be nice to cite that work. You could also cite Matplotlib and Numpy at the very least.
- I also noticed that the image links in the paper are 404'ing. It looks like they broke in your refactor to readthedocs. Could you please get those working too?

### Documentation

I realize I said this:

> You might consider trimming the readme a bit

but I noticed the readme is now much shorter and that you've moved a lot of the information into your readthedocs site. That's great but JOSS only archives the contents of the GitHub repo so a lot of the information you moved to readthedocs should still be in the readme to support long-term preservation. Specifically, JOSS' submission requirements require a README that is closer to how things were before with sections such as:

- A statement of need
- Installation instructions
- Example usage
- Community guidelines (CoC, Contributing, etc)

Could you please move some of the information back in so the readme stands on its own with sections like those I've listed above? Sorry for asking you to duplicate your readme information and switch things back to how they were before.

### Other comments

>  Automated visual regression tests is out of my knowledge.

No worries. I consider this out of scope for our JOSS review. Just an idea.
