### 2018-05-07 #268 @amoeba
Looking good! I made some comments [over there](https://github.com/python-windrose/windrose/issues/77#issuecomment-386989721) and still have a few remaining issues for the author to tackle.
