### 2018-05-08 python-windrose/windrose#77 @scls19fr
Thanks @amoeba for your kind suggestions to improve both this paper and this library.
I will try to tackle remaining issues you mentioned in the next coming days.
