### 2018-05-11 4e25194
Add some contributors, references and bibliography
