+++ paper.bib
@@ -1,0 +1,112 @@
+ @misc{ wiki:###,
+   author = "{Wikipedia contributors}",
+   title = "Wind rose --- {W}ikipedia{,} The Free Encyclopedia",
+   year = "2018",
+   url = "https://en.wikipedia.org/w/index.php?title=Windrose&oldid=839157228",
+   note = "[Online; accessed 11-May-2018]"
+ }
+
+@article{doi:10.1109/MCSE.2007.58,
+  author = {Travis E. Oliphant},
+  title = {Python for Scientific Computing},
+  journal = {Computing in Science \& Engineering},
+  volume = {9},
+  number = {3},
+  pages = {10-20},
+  year = {2007},
+  doi = {10.1109/MCSE.2007.58},
+  URL = { 
+    https://aip.scitation.org/doi/abs/10.1109/MCSE.2007.58
+  },
+  eprint = { 
+    https://aip.scitation.org/doi/pdf/10.1109/MCSE.2007.58
+  }
+}
+
+@article{doi:10.1109/MCSE.2011.36,
+  author = {K. Jarrod Millman and Michael Aivazis},
+  title = {Python for Scientists and Engineers},
+  journal = {Computing in Science \& Engineering},
+  volume = {13},
+  number = {2},
+  pages = {9-12},
+  year = {2011},
+  doi = {10.1109/MCSE.2011.36},
+  URL = { 
+    https://aip.scitation.org/doi/abs/10.1109/MCSE.2011.36
+  },
+  eprint = { 
+    https://aip.scitation.org/doi/pdf/10.1109/MCSE.2011.36
+  }
+}
+
+@article{Walt:2011:NAS:1957373.1957466,
+ author = {Walt, Stefan van der and Colbert, S. Chris and Varoquaux, Gael},
+ title = {The NumPy Array: A Structure for Efficient Numerical Computation},
+ journal = {Computing in Science and Engg.},
+ issue_date = {March 2011},
+ volume = {13},
+ number = {2},
+ month = mar,
+ year = {2011},
+ issn = {1521-9615},
+ pages = {22--30},
+ numpages = {9},
+ url = {https://doi.org/10.1109/MCSE.2011.37},
+ doi = {10.1109/MCSE.2011.37},
+ acmid = {1957466},
+ publisher = {IEEE Educational Activities Department},
+ address = {Piscataway, NJ, USA},
+ keywords = {NumPy, Python, Python, NumPy, scientific programming, numerical computations, programming libraries, numerical computations, programming libraries, scientific programming},
+} 
+
+@article{doi:10.1109/MCSE.2007.55,
+  author = {John D. Hunter},
+  title = {Matplotlib: A 2D Graphics Environment},
+  journal = {Computing in Science \& Engineering},
+  volume = {9},
+  number = {3},
+  pages = {90-95},
+  year = {2007},
+  doi = {10.1109/MCSE.2007.55},
+  URL = {
+    https://aip.scitation.org/doi/abs/10.1109/MCSE.2007.55
+  },
+  eprint = { 
+    https://aip.scitation.org/doi/pdf/10.1109/MCSE.2007.55
+  }
+}
+
+@InProceedings{ mckinney-proc-scipy-2010,
+  author    = { Wes McKinney },
+  title     = { Data Structures for Statistical Computing in Python },
+  booktitle = { Proceedings of the 9th Python in Science Conference },
+  pages     = { 51 - 56 },
+  year      = { 2010 },
+  editor    = { St\'efan van der Walt and Jarrod Millman }
+}
+
+@Misc{,
+  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
[…]
