### 2018-05-11 8bbb51c
modernize Travis-CI
