### 2018-05-11 91a5705
make coding standards a hard failure
