### 2018-05-11 a92a7e9
Work on bibliography/reference
