+++ paper.bib
@@ -41,23 +41,23 @@
 }
 
 @article{Walt:2011:NAS:1957373.1957466,
- author = {Walt, Stefan van der and Colbert, S. Chris and Varoquaux, Gael},
- title = {The NumPy Array: A Structure for Efficient Numerical Computation},
- journal = {Computing in Science and Engg.},
- issue_date = {March 2011},
- volume = {13},
- number = {2},
- month = mar,
- year = {2011},
- issn = {1521-9615},
- pages = {22--30},
- numpages = {9},
- url = {https://doi.org/10.1109/MCSE.2011.37},
- doi = {10.1109/MCSE.2011.37},
- acmid = {1957466},
- publisher = {IEEE Educational Activities Department},
- address = {Piscataway, NJ, USA},
- keywords = {NumPy, Python, Python, NumPy, scientific programming, numerical computations, programming libraries, numerical computations, programming libraries, scientific programming},
+  author = {Walt, Stefan van der and Colbert, S. Chris and Varoquaux, Gael},
+  title = {The NumPy Array: A Structure for Efficient Numerical Computation},
+  journal = {Computing in Science and Engg.},
+  issue_date = {March 2011},
+  volume = {13},
+  number = {2},
+  month = mar,
+  year = {2011},
+  issn = {1521-9615},
+  pages = {22--30},
+  numpages = {9},
+  url = {https://doi.org/10.1109/MCSE.2011.37},
+  doi = {10.1109/MCSE.2011.37},
+  acmid = {1957466},
+  publisher = {IEEE Educational Activities Department},
+  address = {Piscataway, NJ, USA},
+  keywords = {NumPy, Python, Python, NumPy, scientific programming, numerical computations, programming libraries, numerical computations, programming libraries, scientific programming},
 } 
 
 @article{doi:10.1109/MCSE.2007.55,
@@ -109,4 +109,46 @@
   eprint = { 
     https://aip.scitation.org/doi/pdf/10.1109/MCSE.2007.53
   }
+}
+
+@Misc{,
+  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
+  title =     {{SciPy}: Open source scientific tools for {Python}},
+  year =      {2001--},
+  url = "http://www.scipy.org/",
+  note = {[Online; accessed <today>]}
+}
+
+@book{oliphant2006guide,
+  title={A guide to NumPy},
+  author={Oliphant, Travis E},
+  volume={1},
+  year={2006},
+  publisher={Trelgol Publishing USA}
+}
+
+@article{munn1969pollution,
+  title={Pollution wind-rose analysis},
+  author={Munn, RE},
+  journal={Atmosphere},
+  volume={7},
+  number={3},
+  pages={97--105},
+  year={1969},
+  publisher={Taylor \& Francis}
+}
+
+@Misc{,
+  author =    {NRCS National Water and Climate Center},
+  title =     {NRCS National Water and Climate Center - Climate Products - Climate Data - Wind Rose.},
+  url = "http://www.wcc.nrcs.usda.gov/climate/windrose.html",
+  note = {[Online; accessed <today>]}
+}
+
+@Misc{,
+  author =    {Daniel Garver and Ryan Brown},
+  title =     {Air Quality Data Analysis using Open Source Tools - United States Environmental Protection Agency | US EPA},
+  url = "http://slideplayer.com/slide/5922455/",
+  year =      {2016--},
+  note = {[Online; accessed <today>]}
 }
