+++ paper.bib
@@ -1,10 +1,10 @@
- @misc{ wiki:###,
-   author = "{Wikipedia contributors}",
-   title = "Wind rose --- {W}ikipedia{,} The Free Encyclopedia",
-   year = "2018",
-   url = "https://en.wikipedia.org/w/index.php?title=Windrose&oldid=839157228",
-   note = "[Online; accessed 11-May-2018]"
- }
+@misc{wiki:xxx,
+  author = "{Wikipedia contributors}",
+  title = "Wind rose --- {Wikipedia}{,} The Free Encyclopedia",
+  year = "2018",
+  url = "https://en.wikipedia.org/w/index.php?title=Wind_rose&oldid=839157228",
+  note = {[Online; accessed <today>]}
+}
 
 @article{doi:10.1109/MCSE.2007.58,
   author = {Travis E. Oliphant},
