### 2018-05-11 8dcf1cf
Modernize Travis-CI (#86)
* use pytest instead of nose

* add pytest cache dir

* modernize Travis-CI

* make coding standards a hard failure
