+++ paper.bib
@@ -152,3 +152,35 @@
   year =      {2016--},
   note = {[Online; accessed <today>]}
 }
+
+@inproceedings{quick2017optimization,
+  title={Optimization under uncertainty for wake steering strategies},
+  author={Quick, Julian and Annoni, Jennifer and King, Ryan and Dykes, Katherine and Fleming, Paul and Ning, Andrew},
+  booktitle={Journal of Physics: Conference Series},
+  volume={854},
+  number={1},
+  pages={012036},
+  year={2017},
+  organization={IOP Publishing}
+}
+
+@article{harris2014parent,
+  title={The parent wind speed distribution: Why Weibull?},
+  author={Harris, R Ian and Cook, Nicholas J},
+  journal={Journal of wind engineering and industrial aerodynamics},
+  volume={131},
+  pages={72--87},
+  year={2014},
+  publisher={Elsevier}
+}
+
+@article{horel2016summer,
+  title={Summer ozone concentrations in the vicinity of the Great Salt Lake},
+  author={Horel, John and Crosman, Erik and Jacques, Alexander and Blaylock, Brian and Arens, Seth and Long, Ansley and Sohl, John and Martin, Randal},
+  journal={Atmospheric Science Letters},
+  volume={17},
+  number={9},
+  pages={480--486},
+  year={2016},
+  publisher={Wiley Online Library}
+}
