+++ setup.py
@@ -89,8 +89,8 @@ setup(
     # You can install these using the following syntax, for example:
     # $ pip install -e .[dev,test]
     extras_require = {
-        'dev': ['check-manifest', 'nose'],
-        'test': ['coverage', 'nose'],
+        'dev': ['check-manifest', 'pytest'],
+        'test': ['coverage', 'pytest'],
         'advanced': ['pandas', 'scipy']
     },
 
