### 2018-05-11 99a2f63
remove all mentions of nose (#90)
