### 2018-06-13 c169683
Pass Nbins from wrpdf to WindAxes.pdf (#93)
