### 2018-06-16 #268 @scls19fr
This end of year is very difficult and this article is (unfortunately) not on top of my ToDo list. I will tackle this asap but I think you can tag it as "paused" for the next 3 weeks. Sorry and thanks again @amoeba for helping to improve quality.
