+++ CHANGELOG.md
@@ -1,0 +1,56 @@
+# Change Log
+
+## Python Windrose
+
+All notable changes to this code base will be documented in this file,
+in every released version.
+
+### Version 1.6.x
+- Released: 201x-xx-xx
+- Issues/Enhancements:
+  - Improve doc
+  - ...
+
+### Version 1.6.3
+- Released: 2017-08-22
+- Issues/Enhancements:
+  - Issue #69 (PR #70). Dual licensing
+  - ...
+
+### Version 1.6.2
+- Released: 2017-08-02
+- Issues/Enhancements:
+  - Issue #65 (PR #69). Fix inconsistent licence files
+  - ...
+
+### Version 1.6.1
+- Released: 2017-07-30
+- Maintainer: Sébastien Celles <s.celles@gmail.com>
+- Issues/Enhancements:
+  - Original code forked from http://youarealegend.blogspot.fr/search/label/windrose
+  - Create a pip instalable package (registered)
+  - ...
+
+### Version 1.5.0
+- Initial development: 2015-06-16
+- Co-Authors:
+  - Sébastien Celles <s.celles@gmail.com>
+  - Lionel Roubeyrie <lionel.roubeyrie@gmail.com>
+- Maintainer: Sébastien Celles <s.celles@gmail.com>
+- Issues/Enhancements:
+  - Github repository creation
+  - Create a package
+  - Add unit tests / continuous integration
+  - Pandas DataFrames / Series as input (keeping Numpy Array compatibility)
+  - Python 2.7/3.x support
+  - PDF output
+  - Animate windrose (video output)
+  - Add unit tests / continuous integration
+  - PEP8
+  - Register projection windrose
+  - Weitbul plot
+  
+
+### Version 1.4.0
+- Author: Lionel Roubeyrie <lionel.roubeyrie@gmail.com>
+- URL: http://youarealegend.blogspot.fr/search/label/windrose
