### 2018-08-20 6f63b56
Create CONTRIBUTING.md
Closes #89
