### 2018-08-20 29f60cd
Update README.md
Moved to [CHANGELOG(https://github.com/python-windrose/windrose/blob/master/CHANGELOG.md)
