### 2018-08-20 b23303f
Update CONTRIBUTING.md
