### 2018-08-20 python-windrose/windrose#77 @scls19fr
I've just add a [CHANGELOG.md](https://github.com/python-windrose/windrose/blob/master/CHANGELOG.md), [CODE_OF_CONDUCT.md](https://github.com/python-windrose/windrose/blob/master/CODE_OF_CONDUCT.md) and [CONTRIBUTING.md](https://github.com/python-windrose/windrose/blob/master/CONTRIBUTING.md)

I will tackle remaining problems however I don't understand:

> but I noticed the readme is now much shorter and that you've moved a lot of the information into your readthedocs site. That's great but JOSS only archives the contents of the GitHub repo so a lot of the information you moved to readthedocs should still be in the readme to support long-term preservation.

Doc is available on the readthedocs site, but it's still available in docs directory of this Github repository so if "JOSS only archives the contents of the GitHub repo", doc will be also in this archive.
But I can copy a part of content of doc and convert it from RST (docs) to Markdown (for README.md)

About JOSS paper, maybe can I use feedback of some users https://github.com/python-windrose/windrose/issues/54 ?

I will adress version when all this will be ok

Sorry @amoeba for being so late! Python development wasn't my current top priority.
