### 2018-08-20 6c91e2c
Improve paper
