+++ paper/paper.bib
@@ -86,14 +86,6 @@
   editor    = { St\'efan van der Walt and Jarrod Millman }
 }
 
-@Misc{,
-  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
-  title =     {{SciPy}: Open source scientific tools for {Python}},
-  year =      {2001--},
-  url = "http://www.scipy.org/",
-  note = {[Online; accessed <today>]}
-}
-
 @article{doi:10.1109/MCSE.2007.53,
   author = {Fernando Pérez and Brian E. Granger},
   title = {IPython: A System for Interactive Scientific Computing},
@@ -111,7 +103,7 @@
   }
 }
 
-@Misc{,
+@Misc{oliphant2001scipy,
   author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
   title =     {{SciPy}: Open source scientific tools for {Python}},
   year =      {2001--},
@@ -138,14 +130,14 @@
   publisher={Taylor \& Francis}
 }
 
-@Misc{,
+@Misc{nrcs,
   author =    {NRCS National Water and Climate Center},
   title =     {NRCS National Water and Climate Center - Climate Products - Climate Data - Wind Rose.},
   url = "http://www.wcc.nrcs.usda.gov/climate/windrose.html",
   note = {[Online; accessed <today>]}
 }
 
-@Misc{,
+@Misc{garver2016,
   author =    {Daniel Garver and Ryan Brown},
   title =     {Air Quality Data Analysis using Open Source Tools - United States Environmental Protection Agency | US EPA},
   url = "http://slideplayer.com/slide/5922455/",
