### 2018-08-20 34efc85
Update paper.md
Shorten title
