### 2018-08-20 python-windrose/windrose#77 @amoeba
No worries all around.

Re: the README vs. readthedocs: I'm not sure what I meant by that, sorry for the confusion. Maybe I was confused!

As for the README, I interpret the [Documentation](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation) section of the reviewer guidelines to mean that there should be a fair bit of content in the README at the root of the repo (specifically: Statement of need, Installation instructions, Example usage, API documentation, Community guidelines). Some of that stuff is in the `/docs` but if it were moved into the README I think it'd be a bit more visible. I'm okay with content being duplicated between the two. Let me know if you think that's fair.

Scanning the latest version of the paper, it looks really good. However, I think most of that content should be moved over to the README. The [JOSS paper](https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper) is supposed to be fairly short (~1pg) page and:

> Note the paper should not include software documentation such as API (Application Programming Interface) functionality, as this should be outlined in the software documentation.

Sorry if was confusing about where I wanted what. You might take a look at the [reviewer guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html) I'm working off of just to see where I'm coming from.
