### 2018-08-20 1774c28
Create CONTRIBUTORS.md
