+++ README.md
@@ -14,8 +14,92 @@
 
 # Windrose
 
-A windrose, also known as a polar rose plot, is a special diagram for representing the distribution of meteorological datas, typically wind speeds by class and direction.
-This is a simple module for the matplotlib python library, which requires numpy for internal computation.
+A [wind rose](https://en.wikipedia.org/wiki/Wind_rose) is a graphic tool used by meteorologists to give a succinct view of how wind speed and direction are typically distributed at a particular location. It can also be used to describe air quality pollution sources. The wind rose tool uses Matplotlib as a backend. Data can be passed to the package using Numpy arrays or a Pandas DataFrame.
+
+Windrose is a Python library to manage wind data, draw windroses (also known as polar rose plots), and fit Weibull probability density functions.
+
+The initial use case of this library was for a technical report concerning pollution exposure and wind distributions analyzes. Data from local pollution measures and meteorologic informations from various sources like Meteo-France were used to generate a pollution source wind rose.
+
+It is also used by some contributors for teaching purpose 
+![Map overlay](screenshots/overlay.png)
+
+Some others contributors have used it to make figures for a [wind power plant control optimization study](https://www.nrel.gov/docs/fy17osti/68185.pdf).
+
+Some academics use it to track lightning strikes during high intensity storms. They are using it to visualize the motion of storms based on the relative position of the lightning from one strike to the next.
+
+## Install
+
+### Requirements
+
+- matplotlib http://matplotlib.org/
+- numpy http://www.numpy.org/
+- and naturally python https://www.python.org/ :-P
+
+Option libraries:
+
+- Pandas http://pandas.pydata.org/ (to feed plot functions easily)
+- Scipy http://www.scipy.org/ (to fit data with Weibull distribution)
+- ffmpeg https://www.ffmpeg.org/ (to output video)
+- click http://click.pocoo.org/ (for command line interface tools)
+
+### Install latest release version via pip
+
+A package is available and can be downloaded from PyPi and installed using:
+
+```bash
+$ pip install windrose
+```
+
+### Install latest development version
+
+```bash
+$ pip install git+https://github.com/python-windrose/windrose
+```
+
+or
+
+```bash
+$ git clone https://github.com/python-windrose/windrose
+$ python setup.py install
+```
+
+## Examples
+
+The bar plot wind rose is the most common plot
+
+-![Windrose (bar) example](docs/screenshots/bar.png)
+
+Contour plots are also possible
+
+-![Windrose (contourf-contour) example](docs/screenshots/contourf-contour.png)
+
+Several windroses can be plotted using subplots to provide a plot per year with for example subplots per month
+
+-![Windrose subplots](docs/screenshots/subplots.png)
+
+Probability density functions may be plotted. Fitting Weibull distribution is enabled by Scipy.
+The Weibull distribution is used in weather forecasting and the wind power industry to describe wind speed distributions, as the natural distribution of wind speeds often matches the Weibull shape
+
+-![pdf example](docs/screenshots/pdf.png)
 
 ## Documentation
 [![Documentation Status](https://readthedocs.org/projects/windrose/badge/?version=latest)](http://windrose.readthedocs.io/en/latest/?badge=latest)
+Full documentation of library is available at http://windrose.readthedocs.io/
+
+## Community guidelines
+
+You can help to develop this library.
+
+### Code of Conduct
+
+If you are using Python Windrose and want to interact with developers, others users...
+we encourage you to follow our [code of conduct](https://github.com/python-windrose/windrose/blob/master/CODE_OF_CONDUCT.md).
+
+### Contributing
+
+If you discover issues, have ideas for improvements or new features, please report them.
[…]
