### 2018-08-20 32f7928
Minor paper rendering fix
