+++ paper/paper.md
@@ -36,7 +36,7 @@ The initial use case of this library was for a technical report concerning pollu
 
 It is also used by some contributors for teaching purpose.
 
-![Map overlay](screenshots/overlay.png)
+-![Map overlay](screenshots/overlay.png)
 
 Some others contributors have used it to make figures for a [wind power plant control optimization study](https://www.nrel.gov/docs/fy17osti/68185.pdf).
 
@@ -83,20 +83,20 @@ $ python setup.py install
 
 - The bar plot wind rose is the most common plot
 
-![Windrose (bar) example](screenshots/bar.png)
+-![Windrose (bar) example](screenshots/bar.png)
 
 - Contour plots are also possible
 
-![Windrose (contourf-contour) example](screenshots/contourf-contour.png)
+-![Windrose (contourf-contour) example](screenshots/contourf-contour.png)
 
 - Several windroses can be plotted using subplots to provide a plot per year with for example subplots per month
 
-![Windrose subplots](screenshots/subplots.png)
+-![Windrose subplots](screenshots/subplots.png)
 
 - Probability density functions may be plotted. Fitting Weibull distribution is enabled by Scipy.
 The Weibull distribution is used in weather forecasting and the wind power industry to describe wind speed distributions, as the natural distribution of wind speeds often matches the Weibull shape
 
-![pdf example](screenshots/pdf.png)
+-![pdf example](screenshots/pdf.png)
 
 # More advanced usages
 
