### 2018-08-20 python-windrose/windrose#77 @scls19fr
I've updated README and fix some rendering issues in pdf (because of figures).
Sorry for noise because of number of compilation processes.
