+++ CONTRIBUTORS.md
@@ -3,8 +3,8 @@
 - Brian Blaylock - University of Utah Department of Atmospheric Sciences - https://github.com/blaylockbk
 - Ryan Brown - United States Environmental Protection Agency - https://www.epa.gov/
 - daniclaar - Baum lab Applied ecology for impacted oceans - University of Victoria, BC, Canada - https://github.com/daniclaar
-- Filipe Fernandes - Research Software Engineer contractor for SECOORA/IOOS - https://github.com/ocefpaf
 - Sébastien Celles - Université de Poitiers - IUT de Poitiers - https://github.com/scls19fr - project maintainer and co-author
+- Filipe Fernandes - Research Software Engineer contractor for SECOORA/IOOS - https://github.com/ocefpaf
 - Daniel Garver - United States Environmental Protection Agency - https://www.epa.gov/
 - Fabien Maussion - Research Centre for Climate at the University of Innsbruck - https://github.com/fmaussion
 - James McCann - Ramboll - https://github.com/mccannjb
