### 2018-08-22 python-windrose/windrose#77 @scls19fr
I've created a new release (version 1.6.4) on Github https://github.com/python-windrose/windrose/releases
It's also published on Pypi https://pypi.org/project/windrose/
This new release is displayed on Zenodo at https://zenodo.org/record/1401705#.W30PIJP-j_8
But https://github.com/openjournals/joss-reviews/issues/268 is mentioning 1.6 version, it should be 1.6.4 now
Is there anything that I can do ?
