### 2018-08-23 python-windrose/windrose#77 @arfon
> But openjournals/joss-reviews#268 is mentioning 1.6 version, it should be 1.6.4 now
Is there anything that I can do ?

I can update the version on the [review issue](https://github.com/openjournals/joss-reviews/issues/268)
