### 2018-08-23 python-windrose/windrose#77 @scls19fr
Thanks @arfon
