### 2018-08-23 python-windrose/windrose#77 @scls19fr
I wonder why references in https://github.com/python-windrose/windrose/blob/master/paper/paper.bib doesn't appear in https://github.com/openjournals/joss-papers/blob/joss.00268/joss.00268/10.21105.joss.00268.pdf
Should I add references manually in `paper.md`?
