### 2018-08-23 python-windrose/windrose#77 @arfon
This is because you're not cite them directly. (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)). Citations are only included in the PDF if they are cited in the `paper.md`.
