### 2018-08-23 302107a
Update paper.bib
