+++ paper/paper.bib
@@ -77,7 +77,7 @@
   }
 }
 
-@InProceedings{ mckinney-proc-scipy-2010,
+@InProceedings{mckinney-proc-scipy-2010,
   author    = { Wes McKinney },
   title     = { Data Structures for Statistical Computing in Python },
   booktitle = { Proceedings of the 9th Python in Science Conference },
