### 2018-08-24 python-windrose/windrose#77 @amoeba
Looks pretty good, thanks for taking the time to do that. I see a few issues:

README:

- The Codacy badge image is broken
- There's a broken image link in the text "Map overlay" to https://github.com/python-windrose/windrose/blob/master/screenshots/overlay.png
- Could you put code with the examples? I interpret the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), "The authors should include examples of how to use the software (ideally to solve real-world analysis problems)." to mean that there should be some code in the readme with the examples. I find this super helpful when evaluating a library based on the readme and think it would improve this readme.

Paper:

- The paper still contains too much information (See the [criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper)). It looks like you should just be able to remove some material and turn the Examples section into prose.

Let me know what you think and when you want me to take another look. Thanks!
