### 2018-08-27 479c2e3
Update CONTRIBUTING.md
