### 2018-08-27 a93c922
Update paper.md
