### 2018-08-27 python-windrose/windrose#77 @scls19fr
I've fixed `README.md` badges and image (map overlay).
In addition to the documentation, an example have been added to README.
Content of `paper.md` have been reduced.
I will be very pleased if you could you tell me if you think that it matches now criteria?
