### 2018-08-27 python-windrose/windrose#77 @amoeba
hey @scls19fr, thanks for taking the time to do that. The changes the readme are great. It looks like you should remove the Installation section? If you look at an [example paper](https://joss.theoj.org/papers/ba29665c4bff35c37c0ef68cfe356e44), you can see the minimal treatment given to installation instructions. I'd group detailed Installation instructions under the note:

> Note the paper should not include software documentation such as API (Application Programming Interface) functionality, as this should be outlined in the software documentation.
