### 2018-08-27 python-windrose/windrose#77 @amoeba
Great, the revised structure looks good. I just went back to the original review and saw that the references needed checking. My apologies for not looking at that earlier.

- there's extra text 'nocite' in the references section. Does that need to be there?
- `harris2014parent` is missing its DOI of https://doi.org/10.1016/j.jweia.2014.05.005
- `horel2016summer ` is missing its DOI of https://doi.org/10.1002/asl.680
- `munn1969pollution ` is missing its DOI of https://doi.org/10.1080/00046973.1969.9676573

Everything else looks good.
