+++ paper/paper.bib
@@ -121,6 +121,7 @@
 
 @article{munn1969pollution,
   title={Pollution wind-rose analysis},
+  doi = {10.1080/00046973.1969.9676573},
   author={Munn, RE},
   journal={Atmosphere},
   volume={7},
@@ -159,6 +160,7 @@
 @article{harris2014parent,
   title={The parent wind speed distribution: Why Weibull?},
   author={Harris, R Ian and Cook, Nicholas J},
+  doi = {10.1016/j.jweia.2014.05.005},
   journal={Journal of wind engineering and industrial aerodynamics},
   volume={131},
   pages={72--87},
@@ -168,6 +170,7 @@
 
 @article{horel2016summer,
   title={Summer ozone concentrations in the vicinity of the Great Salt Lake},
+  doi = {10.1002/asl.680},
   author={Horel, John and Crosman, Erik and Jacques, Alexander and Blaylock, Brian and Arens, Seth and Long, Ansley and Sohl, John and Martin, Randal},
   journal={Atmospheric Science Letters},
   volume={17},
