### 2018-08-27 python-windrose/windrose#77 @scls19fr
DOI added to `paper.bib` but I also noticed this 'nocite' in the references section.
I followed suggestion from https://github.com/python-windrose/windrose/issues/77#issuecomment-415372276 (ie https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#unused_references_(nocite) ).
Maybe @arfon can tell us why metadata are rendered as raw text?
It can be either a [pandoc](https://pandoc.org/) issue or a misunderstanding from my side of documentation.
