### 2018-08-27 python-windrose/windrose#77 @amoeba
Ah, thanks for the clarification. Let's see what @arfon says.
