### 2018-08-27 python-windrose/windrose#77 @arfon
> Maybe @arfon can tell us why metadata are rendered as raw text?

I think that this section needs to go in the YAML header at the top of the paper.

```
nocite: |
  @wiki:xxx, @doi:10.1109/MCSE.2007.58, @doi:10.1109/MCSE.2011.36, @Walt:2011:NAS:1957373.1957466, @doi:10.1109/MCSE.2007.55, @mckinney-proc-scipy-2010, @doi:10.1109/MCSE.2007.53, @oliphant2001scipy, @oliphant2006guide, @munn1969pollution, @nrcs, @garver2016, @quick2017optimization, @harris2014parent, @horel2016summer
```
