### 2018-08-28 02fdd68
Update paper.md
