### 2018-08-28 python-windrose/windrose#77 @scls19fr
I've just move this section to YAML header at the top of the paper. Rendering looks good now. Thanks @arfon and @amoeba for your help.
