### 2018-08-28 a951dea
fix typo in paper.md (#100)
