### 2018-08-29 cc12c36
Update windrose.py
