### 2018-08-30 c4c1730
Update windrose.py
Closes #99
