### 2018-08-30 #268 @scls19fr
LGTM but let's wait @amoeba opinion.
