### 2018-08-30 #268 @amoeba
Content looks good, thanks @scls19fr. I'm not sure if this was happening before but the PDF version of the paper doesn't show the figures in-line so the examples section is confusing. Could you either find a solution to that or change the bulleted list to a paragraph of text with parenthetical figure references? I'm about to be out of range through Monday in the USA so apologies in advance if my response time drops to zero for a few days here.
