### 2018-08-30 #268 @scls19fr
This change have been done by a contributor in https://github.com/python-windrose/windrose/pull/100
Will revert it. But it adds `-`.
