### 2018-08-30 bbbeee1
Update paper.md
