### 2018-08-30 #268 @scls19fr
Not sure if we can include figure at a specific position without seeing this `-`.
