### 2018-08-30 #268 @amoeba
LGTM, thanks!

@arfon: Review is complete and I approve the submission assuming the version discrepancy (1.6 vs latest) be resolved at the time of archive.

Thanks for listening to my feedback and making changes in a timely manner @scls19fr!
