### 2018-08-30 #268 @amoeba
> Not sure if we can include figure at a specific position without seeing this -.

Oh wow, I didn't even notice that and thought it was just part of the images. I'm not super worried about it at this point.
