+++ windrose/version.py
@@ -1,7 +1,7 @@
 __author__ = 'Lionel Roubeyrie'
 __credits__ = ['Sebastien Celles']
 __email__ = 's.celles@gmail.com'
-__version__ = '1.6.4'
+__version__ = '1.6.5'
 __license__ = 'CeCILL-B OR BSD-3-Clause'
 __url__ = 'https://github.com/python-windrose/windrose'
 __maintainer__ = "Sebastien Celles"
