### 2018-08-30 #268 @scls19fr
v1.6.5 published

See:
- https://zenodo.org/account/settings/github/repository/python-windrose/windrose#
- https://pypi.org/project/windrose/
- https://github.com/python-windrose/windrose

However I can't edit first thread from

    Version: 1.6

to

    Version: 1.6.5
